requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
requests>=2.31.0
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Data sources
pyinaturalist>=0.18.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Computer vision
opencv-python>=4.8.0
scikit-image>=0.21.0

# Machine learning
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0

# Utilities
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
